pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netauto"
version = "1.2.0"
description = "Multi-vendor network automation: config backup, compliance auditing, bulk provisioning"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Network Engineering" }]
keywords = ["networking", "automation", "netmiko", "napalm", "ansible", "bgp", "isis", "mpls"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
]

dependencies = [
    "netmiko>=4.3.0",
    "napalm>=5.0.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "jinja2>=3.1.3",
    "pydantic>=2.6.0",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-PyYAML",
]

[project.scripts]
netauto = "netauto.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["netauto*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "integration: integration tests requiring lab devices",
    "slow: slow-running tests",
]

[tool.coverage.run]
source = ["netauto"]
omit = ["*/tests/*", "*/cli/*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RET"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true
